👌 Extract sessionStoreOperations in its own file
