+++ packages/core/src/domain/session/sessionStoreManager.ts
@@ -1,13 +1,12 @@
-import { clearInterval, setInterval, setTimeout } from '../../tools/timer'
+import { clearInterval, setInterval } from '../../tools/timer'
 import { Observable } from '../../tools/observable'
 import { ONE_SECOND, dateNow } from '../../tools/utils/timeUtils'
 import { throttle } from '../../tools/utils/functionUtils'
 import { generateUUID } from '../../tools/utils/stringUtils'
-import { isChromium } from '../../tools/utils/browserDetection'
-import { SESSION_EXPIRATION_DELAY, SESSION_TIME_OUT_DELAY } from './sessionConstants'
+import { SESSION_TIME_OUT_DELAY } from './sessionConstants'
 import { initCookieStore } from './sessionCookieStore'
-import type { SessionState, SessionStore, StoreInitOptions } from './sessionStore'
-import { isSessionInExpiredState } from './sessionStore'
+import type { SessionState, StoreInitOptions } from './sessionStore'
+import { processSessionStoreOperations } from './sessionStoreOperations'
 
 export interface SessionStoreManager {
   expandOrRenewSession: () => void
@@ -157,108 +156,5 @@ export function startSessionStoreManager<TrackingType extends string>(
     stop: () => {
       clearInterval(watchSessionTimeoutId)
     },
-  }
-}
-
-type Operations = {
-  process: (sessionState: SessionState) => SessionState | undefined
-  after?: (sessionState: SessionState) => void
-}
-
-export const LOCK_RETRY_DELAY = 10
-export const LOCK_MAX_TRIES = 100
-
-const bufferedOperations: Operations[] = []
-let ongoingOperations: Operations | undefined
-
-export function processSessionStoreOperations(operations: Operations, sessionStore: SessionStore, numberOfRetries = 0) {
-  const { retrieveSession, persistSession, clearSession } = sessionStore
-  const lockEnabled = isLockEnabled()
-
-  if (!ongoingOperations) {
-    ongoingOperations = operations
-  }
-  if (operations !== ongoingOperations) {
-    bufferedOperations.push(operations)
-    return
-  }
-  if (lockEnabled && numberOfRetries >= LOCK_MAX_TRIES) {
-    next(sessionStore)
-    return
-  }
-  let currentLock: string
-  let currentSession = retrieveSession()
-  if (lockEnabled) {
-    // if someone has lock, retry later
-    if (currentSession.lock) {
-      retryLater(operations, sessionStore, numberOfRetries)
-      return
-    }
-    // acquire lock
-    currentLock = generateUUID()
-    currentSession.lock = currentLock
-    persistSession(currentSession)
-    // if lock is not acquired, retry later
-    currentSession = retrieveSession()
-    if (currentSession.lock !== currentLock) {
-      retryLater(operations, sessionStore, numberOfRetries)
-      return
-    }
-  }
-  let processedSession = operations.process(currentSession)
-  if (lockEnabled) {
-    // if lock corrupted after process, retry later
-    currentSession = retrieveSession()
-    if (currentSession.lock !== currentLock!) {
-      retryLater(operations, sessionStore, numberOfRetries)
-      return
-    }
-  }
-  if (processedSession) {
-    if (isSessionInExpiredState(processedSession)) {
-      clearSession()
-    } else {
-      processedSession.expire = String(dateNow() + SESSION_EXPIRATION_DELAY)
-      persistSession(processedSession)
-    }
-  }
-  if (lockEnabled) {
-    // correctly handle lock around expiration would require to handle this case properly at several levels
-    // since we don't have evidence of lock issues around expiration, let's just not do the corruption check for it
-    if (!(processedSession && isSessionInExpiredState(processedSession))) {
-      // if lock corrupted after persist, retry later
-      currentSession = retrieveSession()
-      if (currentSession.lock !== currentLock!) {
-        retryLater(operations, sessionStore, numberOfRetries)
-        return
-      }
-      delete currentSession.lock
-      persistSession(currentSession)
-      processedSession = currentSession
-    }
-  }
-  // call after even if session is not persisted in order to perform operations on
-  // up-to-date session state value => the value could have been modified by another tab
-  operations.after?.(processedSession || currentSession)
-  next(sessionStore)
-}
-
-/**
- * Lock strategy allows mitigating issues due to concurrent access to cookie.
- * This issue concerns only chromium browsers and enabling this on firefox increases cookie write failures.
- */
-export const isLockEnabled = () => isChromium()
-
-function retryLater(operations: Operations, sessionStore: SessionStore, currentNumberOfRetries: number) {
-  setTimeout(() => {
-    processSessionStoreOperations(operations, sessionStore, currentNumberOfRetries + 1)
-  }, LOCK_RETRY_DELAY)
-}
-
-function next(sessionStore: SessionStore) {
-  ongoingOperations = undefined
-  const nextOperations = bufferedOperations.shift()
-  if (nextOperations) {
-    processSessionStoreOperations(nextOperations, sessionStore)
   }
 }
